refactor(core): merge http-status-codes lib into core
