+++ cli/thymian.config.yaml
@@ -1,0 +1,15 @@
+plugins:
+  '@thymian/http-linter':
+    options:
+      rules:
+        - '@thymian/rfc-9110-rules'
+      ruleOptions:
+        rfc9110/server-must-send-www-authenticate-header-for-401-response:
+          checkAllSecured: false
+  '@thymian/openapi':
+    options:
+      filePath: openapi.yaml
+  '@thymian/request-dispatcher': {}
+  '@thymian/sampler': {}
+  '@thymian/cli-reporter': {}
+  '@thymian/format-validator': {}
