+++ core/src/http-status-codes/1xx/informational.phrases.ts
