+++ core/src/http-status-codes/1xx/informational.status-codes.ts
