+++ core/tsconfig.json
@@ -4,9 +4,6 @@
   "include": [],
   "references": [
     {
-      "path": "../libs/http-status-codes"
-    },
-    {
       "path": "./tsconfig.lib.json"
     }
   ]
