+++ core/tsconfig.lib.json
@@ -15,9 +15,5 @@
   },
   "include": ["src/**/*.ts"],
   "exclude": ["dist", "test"],
-  "references": [
-    {
-      "path": "../libs/http-status-codes/tsconfig.lib.json"
-    }
-  ]
+  "references": []
 }
