--- libs/http-status-codes/README.md
@@ -1,7 +1,0 @@
-# http-status-codes
-
-This library was generated with [Nx](https://nx.dev).
-
-## Building
-
-Run `nx build http-status-codes` to build the library.
