--- libs/http-status-codes/package.json
@@ -1,11 +1,0 @@
-{
-  "name": "@thymian/http-status-codes",
-  "version": "0.0.1",
-  "files": [
-    "dist",
-    "!**/*.tsbuildinfo"
-  ],
-  "type": "module",
-  "main": "./dist/index.js",
-  "typings": "./dist/index.d.ts"
-}
