--- libs/http-status-codes/tsconfig.lib.json
@@ -1,18 +1,0 @@
-{
-  "extends": "../../tsconfig.base.json",
-  "compilerOptions": {
-    "baseUrl": ".",
-    "rootDir": "src",
-    "outDir": "dist",
-    "tsBuildInfoFile": "dist/tsconfig.lib.tsbuildinfo",
-    "emitDeclarationOnly": false,
-    "moduleResolution": "nodenext",
-    "forceConsistentCasingInFileNames": true,
-    "importHelpers": true,
-    "noImplicitReturns": true,
-    "noFallthroughCasesInSwitch": true,
-    "types": ["node"]
-  },
-  "include": ["src/**/*.ts"],
-  "references": []
-}
