--- libs/http-status-codes/vitest.config.ts
@@ -1,8 +1,0 @@
-import { defineConfig } from 'vitest/config';
-
-export default defineConfig({
-  test: {
-    include: ['test/**/*.test.ts'],
-    passWithNoTests: true,
-  },
-});
