+++ libs/http-testing/package.json
@@ -25,7 +25,6 @@
     "@thymian/core": "^0",
     "@thymian/http-filter": "^0.0.1",
     "rxjs": "^7.8.2",
-    "ajv": "^8.17.1",
-    "@thymian/http-status-codes": "^0"
+    "ajv": "^8.17.1"
   }
 }
