+++ libs/http-testing/tsconfig.lib.json
@@ -21,9 +21,7 @@
     {
       "path": "../http-filter/tsconfig.lib.json"
     },
-    {
-      "path": "../http-status-codes/tsconfig.lib.json"
-    },
+
     {
       "path": "../../core/tsconfig.lib.json"
     }
