+++ plugins/openapi/tsconfig.json
@@ -4,9 +4,6 @@
   "include": ["src/**/*.ts", "test/**/*.ts"],
   "references": [
     {
-      "path": "../../libs/http-status-codes"
-    },
-    {
       "path": "../../core"
     },
     {
