+++ plugins/openapi/tsconfig.lib.json
@@ -16,9 +16,6 @@
   "include": ["src/**/*.ts"],
   "references": [
     {
-      "path": "../../libs/http-status-codes/tsconfig.lib.json"
-    },
-    {
       "path": "../../core/tsconfig.lib.json"
     },
     {
