+++ tsconfig.json
@@ -13,9 +13,6 @@
       "path": "./plugins/openapi"
     },
     {
-      "path": "./libs/http-status-codes"
-    },
-    {
       "path": "./libs/http-testing"
     },
     {
